+++ docs/changelog.rst
@@ -15,6 +15,9 @@ To update |kitty|, :doc:`follow the instructions <binary>`.
 - Fix a regression in 0.20.0 that sent incorrect bytes for the :kbd:`F1-F4` keys
   in rmkx mode (:iss:`3586`)
 
+- macOS: When the Apple Color Emoji font lacks an emoji glyph search for it in other
+  installed fonts (:iss:`3591`)
+
 
 0.20.3 [2021-05-06]
 ----------------------